updated API and create a logger

+++ src/tasks/tasks.service.ts
@@ -9,6 +9,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -19,12 +20,12 @@ export class TasksService {
 
 
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne({ where: { id } });
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!foundTask) {
       throw new NotFoundException();
@@ -32,23 +33,23 @@ export class TasksService {
     return foundTask;
   }
 
-  async createTask(createTaskDto: CreateTaskDTO): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDTO, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id});
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found!`);
     }
   }
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save()
     return task;
-  }
+  } 
 }
